requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.6
future==0.16.0
googlemaps==3.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.0
numpy==1.15.2
oauthlib==2.1.0
Pillow==5.3.0
python-dotenv==0.9.1
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
